requirements.txt
buildbot==0.8.8
PyYAML==3.10
docker-py==0.2.0
